Separate load data handler and normal select handler
